Changes to the AssetManager interface
Previously the interface identified a resource based on its location on the disk. This turned out to be very difficult to implement when working wiht multiple class loaders, so I instead resorted to using input streams. This means that the asset manager will no longer be able to prevent loading the same resource multiple times, but it makes the code more resilient and easier to use.