Implemented a GUI system concept
GUI systems will extend the IEntitySystem and IComponentLoader interfaces. The load and dispose methods are used to initialize and dispose graphics elements, and rendering is performed per-entity in the update method. A similar setup can be used with the ISystem interface instead of IEntitySystem, if rendering is not entity specific.